权限rbac_menu#link_url不能重复
Former-commit-id: 5a1aa7b9a15abe54cb566bfcb9514a81927814d8 [formerly c4d2be3e69c813a113d758e13baca4e44c29d881]
Former-commit-id: 0e19d73ab29dbdef3f87e5bb9091e03f9e865b8b

+++ faber-admin/src/main/java/com/faber/admin/rest/LogApiController.java
@@ -11,6 +11,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
  */
 @Controller
 @RequestMapping("/api/admin/logApi")
-public class LogApiController extends BaseController<LogApiBiz, LogApi, Integer> {
+public class LogApiController extends BaseController<LogApiBiz, LogApi, Long> {
 
 }
